Change Code To OOP Fix Some Bug
