+++ .gitignore
@@ -3,3 +3,5 @@
 ################################################################################
 
 /config/default_config.ini
+/env
+/__pycache__
