+++ .vs/ProjectSettings.json
@@ -1,3 +1,3 @@
 {
-  "CurrentProjectSetting": null
+  "CurrentProjectSetting": "No Configurations"
 }
