+++ .vs/VSWorkspaceState.json
@@ -1,7 +1,8 @@
 {
   "ExpandedNodes": [
-    ""
+    "",
+    "\\config"
   ],
-  "SelectedNode": "\\bot.py",
+  "SelectedNode": "\\default.py",
   "PreviewInSolutionExplorer": false
 }
