+++ .vs/tasks.vs.json
@@ -1,0 +1,10 @@
+{
+  "version": "0.2.1",
+  "tasks": [
+    {
+      "taskLabel": "task-Rollin-Bot",
+      "appliesTo": "bot.py",
+      "type": "launch"
+    }
+  ]
+}
