+++ config/config.ini
@@ -1,7 +1,10 @@
 [config]
 
 #Edit token here
-Token = nothing_here
+Token = "token-here"
 
 #Edit prefix here
 CommandPrefix = >
+
+#Edit log channel here
+LogChannel = "log-id"
